Check required assemblies if YamlDotNet present
Try to determine if YamlDotNet is loaded, and check if it has
required types. Throw an error if the loaded assemblies do not
contain required types. This should at least offer more transparent
error when loading powershell-yaml on a system that uses an
incompatible YamlDotNet assembly.

If no assembly is detected, load the bundled one.

Fixes #46
Partially fixes #47