Use LoadFile() and resolve dependencies manually
LoadFile() works on both dotnet Framework 47 and dotnet Core. The downside is
that we need to manually resolve dependencies for the custom serializer
assembly on dotnet Core. On dotnet Framework there seems to be different behavior
in terms of assembly resolution, and as long as the required assembly is in the
same folder as the requesting assembly, it loads fine.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>