Fix serialization of PSCustomObject wrapped object
With the switch away from converting PSCustomObjects to dictionaries
we broke the way values that are wrapped in a PSCustomObject are serialized.

This is evident when values of hashtables are generated using Write-Output
which generates a PSCustomObject (Which is not evident in powershell, as
there is a lot of magic in powershell), but is clearly evident when we try
to use an object generated in PowerShell, in C#.

This change looks at the nested type of a PSCustomObject, and unwraps it
if it's not just another PSCustomObject.

Fixes: #157

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>