Fix broken json compatibility
The IDictionary and PSCustomObject type converters have to emit a MAppingStart
and MappingEnd. When we added these calls, we ommited to properly set the flow
style for these markers, which brije JSON compatibility.

This change adds the proper options to those Emit calls and adds tests to ensure
Flow styles are obeyed.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>